requirements.txt
amqp==5.2.0
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
billiard==4.2.1
boto3==1.35.57
botocore==1.35.57
cachetools==5.5.0
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron-descriptor==1.4.5
cryptography==43.0.3
daphne==4.1.2
defusedxml==0.8.0rc2
Django==5.1.2
django-appconf==1.0.6
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-channels==0.7.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-ratelimit==4.1.0
django-timezone-field==7.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
google-api-core==2.22.0
google-api-python-client==2.151.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
jmespath==1.0.1
kombu==5.4.2
msgpack==1.1.0
oauthlib==3.2.2
pillow==11.0.0
prompt_toolkit==3.0.48
proto-plus==1.25.0
protobuf==5.28.3
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.2.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.10.3
service-identity==24.2.0
six==1.16.0
social-auth-core==4.5.4
sqlparse==0.5.1
stripe==11.3.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
zope.interface==7.2
